Add files via upload
Adding some user interaction. The user will close the dialog again at some point, so we add a button to close the dialog and assign an event handler.

Fragments are pure UI reuse artifacts and do not have a controller. The third parameter of the sap.ui.xmlfragment function is optional and allows passing in a reference to a (controller) object. However, the third parameter does not necessarily have to be a controller but can be any object.

The event handler function is put into the same controller file and it closes the dialog by accessing the internal helper function that returns the dialog.